chore: improve Toast exports
